requirements.txt
torch==2.7.*
torchvision
numpy
pillow
pandas
kagglehub
